Language.PlutusTx.Lift

Documentation

makeLift :: Name -> Q [Dec]

lift :: (Lift a, AsError e (Provenance ()), MonadError e m, MonadQuote m) => a -> m (Term TyName Name ())

Get a Plutus Core term corresponding to the given value.

liftProgram :: (Lift a, AsError e (Provenance ()), MonadError e m, MonadQuote m) => a -> m (Program TyName Name ())

Get a Plutus Core program corresponding to the given value.

unsafeLift :: Lift a => a -> Term TyName Name ()

Get a Plutus Core term corresponding to the given value, throwing any errors that occur as exceptions and ignoring fresh names.

unsafeLiftProgram :: Lift a => a -> Program TyName Name ()

Get a Plutus Core program corresponding to the given value, throwing any errors that occur as exceptions and ignoring fresh names.