Language.PlutusTx.Lift.Class

Documentation

class Typeable (a :: k) where

Class for types which have a corresponding Plutus IR type. Instances should usually be declared for type constructors, instances for applied types will be derived.

Minimal complete definition

typeRep

Methods

typeRep :: RTCompiling m => Proxy a -> m (Type TyName ())

Get the Plutus IR type corresponding to this type.

class Lift a where

Class for types which can be lifted into Plutus IR.

Minimal complete definition

lift

Methods

lift :: RTCompiling m => a -> m (Term TyName Name ())

Get a Plutus IR term corresponding to the given value.

makeTypeable :: Name -> Q [Dec]

makeLift :: Name -> Q [Dec]